Rename service to TagEnforcement.
